Restructure NewTransactionView
- Create a folder New Transaction
- Move subviews into their own folder
